refactor: split Java compiler
